FIX: silent failure when watched words contains invalid regex
If a watched words list contains any invalid regex, the test modal won't
work because the "compiled" regex (that contains all the watched words)
will be invalid.

This PR introduces a fallback to matching individual watched words when
the compiled regex is invalid as well as displaying the error for each
invalid regex.

Internal ref - t/168206