FIX: "browse latest topics" button in new new topics / replies subtabs
wasn't working.

The issue was that when we were in a subtab, and there was no more "new"
topics or replies, the "ctaRoute" was shunted because we checked against
"this.args.newListSubset" (which is always truthy in those subtabs).

To fix it, I changed the conditionals to test against the presence of
"this.ctaLabelWithAction.action" so that whenever the CTA button has an
action, we don't use the route. When it doesn't have an action, we use
the route.

This is a follow up of 66179c60cd47e

Internal ref - t/168867