FIX: reactions count when there are deprecated bookmarks
When computing the number of reactions on a post, we were missing a
filter to ensure we were correctly counting likes.

Also improved the performance of the computation of the reactions count
by doing all the work in SQL instead of retrieving all the data, parsing
it in Ruby and filtering it server-side.

Also slightly refactored the specs to use "contain_exactly" instead of
"eq" because the order doesn't matter.

Internal ref - t/168844