FIX: Allow user to choose if they want to edit localized or original title (#36099)
When a user is viewing a localized title, and edits the title, they are
shown the original title.

They are then faced with the conundrum of updating it without being sure
if they are editing the original or localization.

This commit allows them to choose in cases where it is ambiguous.

Meta: https://meta.discourse.org/t/topic-title-translation-issues/388848