FEATURE: Ability to pin chat channels
## :mag: Overview
This update adds the ability to pin chat channels. Pinned channels take priority in the sidebar channel list and as such are shown at the top of the channel list. To configure a channel to be pinned you must configure it in the channel settings. Pinning channels is a user specific configuration and will be applied for the current user rather than site wide.

## 📷 Screenshots