DEV: Remove AI preference tab and move setting to "other" (#36144)
## :mag: Overview
This update removes the AI user preference tab since there is only one
setting present there (conditionally). Instead we move that setting to
appear in the "other" section of the user's interface preferences. We
can add back the AI user preferences page later on if we have many more
settings but at this time it doesn't feel necessary.