FIX: type object setting not redirecting on saving (#36150)
With https://github.com/discourse/discourse/pull/35349, now site
settings do not return a body on update, which broke the type object
setting editor's save flow. This commit adds handling for that case.

And also adds testing to ensure that after saving the setting, we are
redirected to another page.

Renamed `admin_objects_theme_setting_editor` to
`admin_objects_setting_editor` since it can be used for both themes and
settings.


Before:


https://github.com/user-attachments/assets/5e019428-e126-4084-80a6-eb324c851427


Now:


https://github.com/user-attachments/assets/9ac59ca2-1896-490a-866c-630a5a201471

A future todo would be to migrate [ThemeController#update_single_setting
](https://github.com/discourse/discourse/blob/71834c898f2f3f5d11df3db6f9a5bab12acbccaa/app/controllers/admin/themes_controller.rb#L343-L365)
to use the same pattern as
[`SiteSettingController#update`](https://github.com/discourse/discourse/blob/d8e7741d9645d39a637037c7720dd6fc5f261284/app/controllers/admin/site_settings_controller.rb#L25-L53)