FIX: prevents scroll on focus when doing autofocus (#36164)
Our current trap-tab modifier will also attempt to autofocus the first
focusable element. This can cause issues on iOS where the autofocus
behavior is buggy and will attempt to scroll the body even if
`preventScroll: true` is used.

This commit does the following changes:
- applies `focusOffScreen` on iOS when we attempt to autofocus
- renames `positionOffscreen` to `focusOffScreen` as it's more inline
with its behavior
- reworks the `wait-for-closed-keyboard` library to rely on the existing
`keyboard-visible` class added/removed to documentElement, we could have
used appEvents to listen on keyboard-visibility change but that felt
cleaner to use a mutation observer. I noticed our current version can be
very slow on iOS26
- settle on a common `KEYBOARD_DETECT_THRESHOLD` at 150px, which seems
to be the minimum height we could ever encounter