DEV: Improve Pitchfork workers timeouts
Currently, when a Pitchfork worker times out, we’re logging its backtrace
as a warning. After that, Pitchfork will terminate cleanly the worker by
calling `Process.exit`. This, however, raises a `SystemExit` exception
that is caught by Logster. Since we already have the backtrace logged,
we don’t need to log that exception which is more confusing than
anything else.

This change makes Logster ignore that `SystemExit` exception and raises
the log level to error instead of warning.