DEV: Register container element for reactions (#36169)
## Background

When multiple reactions sharing the same post ID are rendered on a
single page, the reactions picker exhibits incorrect targeting behavior.
Instead of responding to the element the user is hovering over, it
consistently targets the first matching element.

## Changes

This pull request introduces a reference to the container element via
the didInsert modifier, allowing the reactions picker to target the
specific element being interacted with, thereby resolving the targeting
conflict.