FIX: Post data on lightbox slide in production
Followup 006b7e0d9f06a7c02501189387e6e2f9e98da756

When calling this code in production:

```
if (relatedModel?.constructor?.name === "Post") {
  data.post = relatedModel;
}
```

The name of the constructor is being minified, so the check fails and
the post data is not set on the lightbox data. The name of the
constructor is something like "F" in production.

Instead, we can just make the data param of the lightbox() init function
more generic, so we can specify the keys we want to set on the data
object.