DEV: Stop logging theme load errors to logster
In our experience, this is creating a significant amount of un-actionable noise in the logs. Themes can fail to boot for all sorts of client-caused reasons, like a poor network connection, adblocker, unsupported browser, etc.

Logging those to Logster is not helpful, and only leads to confusion for admins and hosting providers.