DEV: Introduce new versioning system (#35180)
Implements date-based versioning, based on the RFC at
https://meta.discourse.org/t/383536

Manual version-bump tasks are removed, and replaced with new `release:*`
rake tasks. These are run in GitHub actions. Release process will work
something like:

1. Trigger `release-prepare-latest-bump` via workflow_dispatch. This
will create a PR which bumps the in-development version to the next
`-latest`

2. Merge that PR

3. `release-handler` will be triggered automatically, tag the commit
with `v0000.00.0-latest`, and cut a `release/xxxx.xx` branch from the
previous commit

4. `release-prepare-bump` will be triggered automatically, and create a
PR which bumps the version on `release/xxxx.xx` branch to remove the
`-latest` suffix

5. Merge that PR

6. `release-handler` will be triggered automatically, and will tag the
commit with `v0000.00.0`

In future, we will add handling for ESR, and new workflows for security
fixes.