FIX: translation missing when using optional_* placeholder in body (#36023)
All the "optional_*" placeholders are meant to be used in the subject
but if you were to use them in the body template you would get a
"translation missing" error.

This commit ensures these optional placeholders are also provided when
i18ning the body even though they don't make a lot of sense.

Internal ref - t/150916