A11Y: make user-info component headings optional (#36209)
The `role="heading"` was added to user-info in
https://github.com/discourse/discourse/pull/30499 to improve the user
directory accessibility... but we use this component in a number of
places and they don't all need to be headings (most don't!).

This adds an optional condition to include `headingLevel` and includes
it for the directory. This ensures the level is included along with the
heading role (previously we had no level at all, but it's required
according to
https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Reference/Roles/heading_role#associated_wai-aria_roles_states_and_properties)