A11Y: Hide avatar link from tabindex when it's aria-hidden (#36227)
When paired with a username, we hide the avatar link from assistive
technology because it's a decorative image wrapped in a duplicate link.
We don't change the tabindex though, which means it's still reachable
with tab navigation... and in this situation a screenreader may still
announce it.

<img width="300" alt="image"
src="https://github.com/user-attachments/assets/9ff60992-8956-4b96-89a7-e3c1a5f6ec59"
/>

This adds tabindex `-1` when `ariaHidden` to remove it from tab nav so
it's no longer an issue.