FIX: `redis.setex()` expects seconds parameter not DateTime (#36238)
## :mag: Overview

This update fixes a bug in chat's `set_replies_count_redis_cache`
method.

`Discourse.redis.setex()` expects the 2nd parameter to be a seconds
value not a `DateTime` value as an integer. This was causing thread
cache keys to last for numerous years instead of 5 mintues in Redis. The
value should be 300 seconds (5 minutes) instead of `5.minutes.from_now`
date/time getting converted to an integer.