FIX: allows float-kit to accept a portalOutletElement
By default float kit instance will be rendered in top level dom containers to simplify positioning and z-index. However in some cases this can be an issue. One notable issue is for triggers located in a sticky elements, the positioning of the floating element can lag behind when scrolling fast. In this case the developer can request to render the floating content inside the sticky container to avoid this issue.

This commit has an example of this fix in the user-tip used for the topic-timeline container component.