PERF: Fix memory leak in server-side plugin outlets (#36284)
We've identified a memory leak that occurs when rendering custom
templates into server-side plugin outlets where each rendered template
leaves at least 1 object in the heap that can never be claimed back by
the GC. This happens due to ActionView's caching of compiled template
and is [intended
behavior](https://github.com/rails/rails/issues/33019#issuecomment-409379676),
however in our case, this is problematic because it can make memory
usage grow extremely fast if a plugin connects to an outlet on a page
that gets visited a lot (e.g. the topic page).

This PR moves away from using ActionView's `render` method, which
creates a new `ActionView::Template` instance which gets compiled and
cached, to a custom solution that creates an `ActionView::Template`
instance only once for a given template and reuses the instance when the
template is needed again.

In development, to avoid having to restart the server when changing a
plugin's template, the `PLUGIN_OUTLET_TEMPLATE_CACHE` constant where
`ActionView::Template` are kept, gets reset on code reload via the
`to_prepare` hook.

Internal topic: t/169204.