DEV: allow form fields that are outside of modal (#36312)
Changes in #36268 causes form fields that are rendered outside of the
modal to `preventDefault`. In the case of a text field or textarea it
prevents typing.

I think stopping the event propagation should be enough and it solves
the issue of editing form fields that are rendered outside of modals
(ie. FloatKit etc).