DEV: follow-up fixes for welcome banner admin config area (#36318)
Follow-ups to
https://github.com/discourse/discourse/commit/878d23a1afbb4e819382e9ba30934888ed64d59b

The admin area adds a class based on the route name... previously this
was `.welcome-banner` which was picking up styles from the actual
welcome banner. Changing it to `adminWelcomeBanner` makes it
`.admin-welcome-banner` to avoid this, while the path can remain the
same.

I also noticed some broken placeholder references in the headline
inputs, so now those are passed in correctly.