FIX: Check for pgvector before CREATE EXTENSION (#35070)
On certain hosting environments (azure postgres flex notably) CREATE
EXTENSION requires a high level of permissions if the extension in
question is marked as "untrusted" by postgres. This level of permission
might not be available at runtime if the database user has low
privileges.

Unfortunately, the migration currently tries to CREATE EXTENSION even if
the "vector" extension already exists, which causes an unavoidable
permission error.

This PR just adds a check to see if "vector" is already present, and
skips if it is.