FIX: Related topics causing page render issues (#36357)
## :mag: Overview

`store.createRecord()` was being called inside a getter that was
accessed during render. This was causing issues because
`createRecords()` hydrates topic models setting tracked properties like
`bookmarks`. However, since `bookmarks` was already being consumed in
the same render cycle, it was resulting in page re-rendering issues
because Ember was crashing from autotracking violations.

This update ensures that the the raw data from the server is transformed
as soon as the data arrives instead of during render.

We also remove the `@cached` decorator here as the admonition is to
avoid them to improve performance (see /t/149819).