FIX: Fix migration of HouseAd records from the `plugins-store-rows` table when `visible_to_logged_in_users` and/or `visible_to_anons` options are not set in the DB (#36362)
**Description**

When the post migration to move HouseAd records from the
`plugins-store-rows` is run, it currently expects that the
`visible_to_logged_in_users` and `visible_to_anons` options are set.
However, this is not the case; this PR assumes the lack of these options
as false, allowing the migration to run.