FIX: allow entering 0.002 in number inputs (#36374)
We call parse float and 0.00 will parse to 0 so when you start typing
`0.0` in the field it will right away change it to `0`, which could be
ok to say `0.0` is `0`, but this will prevent you to type anything like
`0.01`.

There is still a slightly annoying case:
- input `0` -> value `0`
- input `0.` -> value `null`

So we have an input but we will report `null` to the form, I think this
is an acceptable risk for now.

Note: the test is not testing correctly behavior, we tried to use
`typeIn` instead of `fillIn` but the input will also not accept `0.` and
prevent us to test it correctly.

---------

Co-authored-by: Joffrey JAFFEUX <j.jaffeux@gmail.com>