FIX: Retranslate topic excerpts and avoid using post versions to indicate change (#36382)
There are three bugs fixed here:

## bug 1
`revisor.should_create_new_version?` does not trigger when an upload is
added. This means localizations will not be updated. This is now fixed
by not using `should_create_new_version` but `raw_changed` instead.

m/t/389685/15

## bug 2
When the topic's first post is updated, it does not update the
topic_localization excerpt. This is now fixed.

t/169370/3

## bug 3
When the topic title is edited (entry_point.rb), the
`detect_translate_topic` job is enqueued, however, localization does not
happen because it already exists. The job now is consistent and
relocalizes based on a daily quota.

t/169370/3