FIX: Migration issues of HouseAd records from the plugin-store-rows table when they reference non existing categories (#36380)
**Description**

When the post migration to move HouseAd records from the
plugins-store-rows is run, it currently expects that the referenced
categories in the `category_ids` option exist. However, this is not the
case. Currently, when a HouseAd record references a category that
doesn't exist, it doesn't show in the UI. Taking this into account, this
PR follows the same behaviour by ignoring those non-existing category
references but creating the HouseAd record.