PERF: Remove unused MessageBus subs to PM unread indicator channels (#36376)
There are no elements that match the `.indicator-topic-${data.topic_id}`
selector (they used to exist), so we can safely stop subscribing to the
MessageBus channel for each topic in the topic list since the only thing
that the callback does is updating the (now) non-existing elements.

Internal topic: t/169204/25.