DEV: Move AI credits daily usage to it's own table (#36369)
## :mag: Overview
This update moves the AI credit daily usage data from being a `jsonb`
value in the `LlmCreditAllocation` table to it's own dedicated table
with each day as it's own record. There is still some of the old code
for now as we migrate existing sites over, which we will eventually
follow-up and remove the deprecated code.