FEATURE: Ability to star chat channels (#36142)
## :mag: Overview
This update adds the ability to star chat channels. Starred channels can
include DM channels or chat channels. Once a channel is starred it will
create a new section in the sidebar that can be collapsed/expanded. It
also adds a new tab to the mobile chat interface. Starring channels is a
user specific configuration and will be applied for the current user
rather than site wide.

At the moment this is enabled as an upcoming pre-alpha change by
toggling the settings for the respective groups in the upcoming changes
page:

<img width="986" height="141" alt="Screenshot 2025-11-28 at 14 45 07"
src="https://github.com/user-attachments/assets/1b297701-9135-4b6f-b36b-0805975a7872"/>

## 📹 Preview

### Desktop

https://github.com/user-attachments/assets/a0f84077-65a1-4f23-95f7-ac575c1387fa

### Mobile

https://github.com/user-attachments/assets/15ce190c-a647-4ae8-b552-c67686be56ea

---------

Co-authored-by: Martin Brennan <martin@discourse.org>