FIX: update rack sendfile middleware for x-accel-redirect (#36354)
Background: mp4 video type was not playing in Safari on macOS or iOS if
served from Discourse instead of S3. The response was a 200 instead of a
206 for partial response. There was a
[change](https://github.com/rack/rack/compare/v2.2.19...v2.2.20) in Rack
v2.2.20 that forces apps to be explicit about X-Accel-Redirect on the
Sendfile command

Change: Be explicit about enabling X-Accel-Redirect in application
middleware for Rack::Sendfile.
Also add guards for environment to not get `X-Accel-Mapping header
missing` on local env logs