DEV: map reactions plugin heart to d-icon aliases (#36467)
In core we have aliases for the liked/unliked icon for posts so that
they can be replaced without globally replacing the heart icon in other
contexts.

We aren't currently using these aliases for the reactions like button,
so like icon replacements will regress when an admin turns on reactions.

This updates the plugin so that when the default "heart" value is set as
`discourse_reactions_like_icon` we use the alias, which will allow
existing customizations of the like icon to continue working with
reactions. Any changes to the `discourse_reactions_like_icon` setting
will still be respected.