FIX: Revert composer-open body style to #36253 to fix reply box flicker on IOS devices (#36449)
In #36331, a fix is implemented to the chat box to make it scrollable.
However, the scss property `html.ios-device.composer-open body` is also
affected.

The new style `position: fixed` makes the page going up unstoppable when
users trying to open reply box, thus causing a flicker.

This PR fix the this by change back the attribution style to `overflow:
hidden` used in #36253.